{{ cookiecutter.repo_name }}/requirements.txt
opencensus-extension-azure-functions
opencensus-ext-requests
azure-functions-core